+++ common/src/main/kotlin/gg/flyte/pluginportal/common/adapters/PolymartAdapter.kt
@@ -1,16 +1,12 @@
 package gg.flyte.pluginportal.common.adapters
 
-import com.google.gson.JsonObject
-import gg.flyte.pluginportal.common.PluginPortalBase
 import gg.flyte.pluginportal.common.types.LocalPlugin
 import gg.flyte.pluginportal.common.types.enums.MarketplacePlatform
-import gg.flyte.pluginportal.common.util.GSON
 import gg.flyte.pluginportal.common.util.HashType
 import gg.flyte.pluginportal.common.util.download
+import gg.flyte.pluginportal.common.util.getPolymartDownloadUrl
 import java.io.File
-import java.net.HttpURLConnection
 import java.net.URL
-import java.net.URLEncoder
 
 class PolymartAdapter : DownloadAdapter {
     override fun getPriority(): Int = 200
@@ -36,7 +32,7 @@ class PolymartAdapter : DownloadAdapter {
         
         // For free plugins, we can use the direct download URL
         try {
-            val downloadUrl = getPolymartDownloadUrl(polymart.platformId, null, isPremium)
+            val downloadUrl = getPolymartDownloadUrl(polymart.platformId)
                 ?: return DownloadResult(false, error = "Failed to get download URL from Polymart API")
             
             // Download the file
@@ -63,54 +59,4 @@ class PolymartAdapter : DownloadAdapter {
             return DownloadResult(false, error = "Download failed: ${e.message}")
         }
     }
-    
-    
-    private fun getPolymartDownloadUrl(resourceId: String, token: String?, isPremium: Boolean): String? {
-        return try {
-            val url = URL("https://api.polymart.org/v1/getDownloadURL")
-            val connection = url.openConnection() as HttpURLConnection
-            connection.requestMethod = "POST"
-            connection.doOutput = true
-            connection.setRequestProperty("Content-Type", "application/x-www-form-urlencoded")
-            
-            val postData = buildString {
-                append("resource_id=${URLEncoder.encode(resourceId, "UTF-8")}")
-                // Only add token parameter if it's not null and not blank
-                // For free resources, we should NOT send any token parameter at all
-                if (!token.isNullOrBlank()) {
-                    // For premium resources, this must be a valid OAuth token
-                    // obtained through the user authorization flow
-                    append("&token=${URLEncoder.encode(token, "UTF-8")}")
-                }
-            }
-
-            PluginPortalBase.plugin.logger.info(
-                "Requesting Polymart download URL from $url for resource $resourceId (${if (isPremium) "premium" else "free"})"
-            )
-            connection.outputStream.use { it.write(postData.toByteArray()) }
-            
-            val responseCode = connection.responseCode
-            
-            if (responseCode == 200) {
-                val response = connection.inputStream.bufferedReader().use { it.readText() }
-                val json = GSON.fromJson(response, JsonObject::class.java)
-                
-                if (json.getAsJsonObject("response")?.get("success")?.asBoolean == true) {
-                    return json.getAsJsonObject("response")
-                        ?.getAsJsonObject("result")
-                        ?.get("url")?.asString
-                } else {
-                    val error = json.getAsJsonObject("response")?.get("message")?.asString
-                    PluginPortalBase.plugin.logger.warning("Polymart download URL request failed: ${error ?: "unknown error"}")
-                }
-            } else {
-                PluginPortalBase.plugin.logger.warning("Polymart download URL request failed with HTTP $responseCode")
-            }
-            null
-        } catch (e: Exception) {
-            PluginPortalBase.plugin.logger.warning("Polymart download URL request failed: ${e.message ?: e::class.simpleName}")
-            null
-        }
-    }
-    
 }
